Map draws from user view now
Changed the way the map is drawn. Instead of drawing whole map starting in top left, the map is drawn starting at the position of the player and then is drawn outward in all directions up to the visibility setting. It is NOT allowed to "see" through a wall (the map drawing stops in a direction when a wall is encountered)